refactor: add forum review remarks

+++ packages/server/api/services/ForumMessage.ts
@@ -17,7 +17,8 @@ export const forumMessageRoute = Router()
       .catch(next);
   })
   .post('/', (req: Request, res: Response, next) => {
-    if (res.locals.user && res.locals.user.id === req.body.user_id) {
+    if (res.locals.user && res.locals.user.id ) {
+      req.body.user_id = res.locals.user.id;
       ForumMessage.create(req.body, { include: [{ model: User }] })
         .then(message => {
           ForumMessage.findByPk(message.id, { include: [{ model: ForumTopic }, { model: User }] })
@@ -30,11 +31,14 @@ export const forumMessageRoute = Router()
     }
   })
   .put('/:id', (req: Request, res: Response, next) => {
-    if (res.locals.user && res.locals.user.id === req.body.user_id) {
-      ForumMessage.update(req.body, { where: { id: req.params.id }, returning: true })
+    if (res.locals.user && res.locals.user.id) {
+      ForumMessage.update(req.body, { where: { id: req.params.id, user_id: res.locals.user.id }, returning: true })
         .then(result => {
-          const [, message] = result;
-          ForumMessage.findByPk(message[0].id, { include: [{ model: ForumTopic }, { model: User }] })
+          const [count, messages] = result;
+          if (count === 0) {
+            throw Error('Сообщение не найдено');
+          }
+          ForumMessage.findByPk(messages[0].id, { include: [{ model: ForumTopic }, { model: User }] })
             .then(message => res.status(200).json(message))
             .catch(next);
         })
@@ -44,7 +48,12 @@ export const forumMessageRoute = Router()
     }
   })
   .delete('/:id', (req: Request, res: Response, next) => {
-    ForumMessage.destroy({ where: { id: req.params.id } })
+    if (res.locals.user && res.locals.user.id) {
+    ForumMessage.destroy({ where: { id: req.params.id, user_id: res.locals.user.id } })
+      .then(throwIf(r => !r, res, 400, 'Комментарий не найден'))
       .then(() => res.status(201).send({ message: 'Комментарий удален' }))
       .catch(next);
+    } else {
+      res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
+    }
   });
