+++ packages/server/api/services/ForumTopic.ts
@@ -34,20 +34,23 @@ export const forumTopicRoute = Router()
       .catch(next);
   })
   .post('/', (req: Request, res: Response, next) => {
-    if (res.locals.user && res.locals.user.id === req.body.user_id) {
-    ForumTopic.create(req.body)
-      .then(topic => res.status(201).send({ id: topic.id }))
+    if (res.locals.user && res.locals.user.id ) {
+      req.body.user_id = res.locals.user.id;
+      ForumTopic.create(req.body).then(topic => res.status(201).send({ id: topic.id }))
      .catch(next);
     } else {
       res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
     }
   })
   .put('/:id', (req: Request, res: Response, next) => {
-    if (res.locals.user && res.locals.user.id === req.body.user_id) {
-    ForumTopic.update(req.body, { where: { id: req.params.id }, returning: true })
+    if (res.locals.user && res.locals.user.id) {
+    ForumTopic.update(req.body, { where: { id: req.params.id, user_id: res.locals.user.id }, returning: true })
       .then(result => {
-        const [, messages] = result;
-        res.status(200).json(messages[0]);
+        const [count, topics] = result;
+        if (count === 0) {
+          throw Error('Топик не найден');
+        }
+        res.status(200).json(topics[0]);
       })
       .catch(next);
     } else {
@@ -55,7 +58,12 @@ export const forumTopicRoute = Router()
     }
   })
   .delete('/:id', (req: Request, res: Response, next) => {
-    ForumTopic.destroy({ where: { id: req.params.id } })
+    if (res.locals.user && res.locals.user.id) {
+    ForumTopic.destroy({ where: { id: req.params.id, user_id: res.locals.user.id } })
+      .then(throwIf(r => !r, res, 400, 'Тема не найдена'))
       .then(topic => res.status(201).json(topic))
       .catch(next);
+    } else {
+      res.status(500).send({ type: 'error', message: 'Доступ запрещен' });
+    }
   });
